fix 1 move piece display bug
Earlier pieces were shown only from the second FEN position for a new
game. This bug was introduced as a regression when I wanted to draw
last move square before draw squares function so that the last move
highlighted square lay beneath the pieces.

Also introduced button array whose buttons will be packed into a compact
display as a vbox to the right of the current window. There will also be
an evaluation bar to the left of the current window.